:truck: Move actions into lib directory
